+++ app/javascript/beyond_canvas/initializers/inputs.js
@@ -1,4 +1,4 @@
-(function($) {
+(function ($) {
   const onDOMReady = function () {
     $('input[type="file"]').each(function () {
       var $input = $(this),
@@ -9,10 +9,7 @@
         var fileName = '';
 
         if (this.files && this.files.length > 1)
-          fileName = (this.getAttribute('data-multiple-caption') || '').replace(
-            '{count}',
-            this.files.length
-          );
+          fileName = (this.getAttribute('data-multiple-caption') || '').replace('{count}', this.files.length);
         else if (e.target.value) fileName = e.target.value.split('\\').pop();
 
         if (fileName)
@@ -33,6 +30,5 @@
     });
   };
 
-  $(document)
-    .on('ready page:load turbolinks:load', onDOMReady);
+  $(document).on('ready page:load turbolinks:load', onDOMReady);
 })(jQuery);